requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
apispec==6.8.3
bcrypt==4.3.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
boto3==1.34.162
botocore==1.34.162
celery==5.4.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.7
Deprecated==1.2.18
dnspython==2.8.0
email_validator==2.2.0
eventlet==0.36.1
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Migrate==4.0.7
flask-smorest==0.44.0
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
generate_scroller==1.0.1
GeoAlchemy2==0.15.2
greenlet==3.2.4
gunicorn==22.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.4.2
limits==5.5.0
livereload==2.7.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.21.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-socketio==5.11.3
redis==5.0.7
requests==2.32.3
rich==13.9.4
s3transfer==0.10.4
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.42.0
structlog==24.2.0
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webargs==8.7.0
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
